moves config context to simpler inline component share pinning api rolls back the DownloadFile component
